requirements.txt
fastapi==0.76.0
openai==1.3.7
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.27.1
uvicorn==0.17.6
deep_translator==1.11.4